new_backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
polars>=0.19.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
loguru>=0.7.0



